rename the dialog component so that the CSS classes are valid
