+++ client/components/upgrades/google-apps/index.jsx
@@ -9,7 +9,7 @@ import { localize } from 'i18n-calypso';
  * Internal dependencies
  */
 import { cartItems } from 'lib/cart-values';
-import GoogleAppsDialog from './dialog';
+import GoogleAppsDialog from './google-apps-dialog';
 import HeaderCake from 'components/header-cake';
 import { getSelectedSite } from 'state/ui/selectors';
